Placed NumberedBloomFilter class into LayeredBloomFilterTest where it is used and fixed implementation
