+++ src/test/java/org/apache/commons/collections4/bloomfilter/LayerManagerTest.java
@@ -289,13 +289,4 @@ public class LayerManagerTest {
         assertEquals(2, supplierCount[0]);
     }
 
-    static class NumberedBloomFilter extends WrappedBloomFilter {
-        int value;
-        int sequence;
-        NumberedBloomFilter(Shape shape, int value, int sequence) {
-            super(new SimpleBloomFilter(shape));
-            this.value = value;
-            this.sequence = sequence;
-        }
-    }
 }
